comment actions toggling is working
