+++ src/components/specific/visa/visa-summary/visa-comments/visa-comment-post/VisaCommentPost.scss
@@ -1,4 +1,4 @@
-.visa-comments-list {
+.visa-comment-post {
   &__header {
     display: flex;
     justify-content: space-between;
@@ -48,10 +48,4 @@
 }
 .reply {
   margin-left: 44px;
-}
-.safe-zone {
-  z-index: 2;
-  position: absolute;
-  top: 2.5%;
-  left: 3.5%;
 }
