+++ src/main.js
@@ -1,15 +1,15 @@
 /* eslint-disable */
 import { createApp } from "vue";
-import vueClickAway from "vue3-click-away";
 import i18n from "@/i18n";
 import router from "@/router/index.js";
 import ErrorService from "@/services/ErrorService";
 
 import App from "./App.vue";
 import globalComponents from "@/components/global-components";
+import { BIMDataClickAway } from "@bimdata/design-system/dist/js/BIMDataDirectives/vue3/index.js";
+
 
 const app = createApp(App)
-  .use(vueClickAway)
   .use(i18n)
   .use(router);
 
@@ -17,6 +17,8 @@ const app = createApp(App)
 for (const [name, component] of Object.entries(globalComponents)) {
   app.component(name, component);
 }
+
+app.directive('clickAway', BIMDataClickAway)
 
 // Setup global error handler
 app.config.errorHandler = error => {
